requirements.txt
Django==1.8.2
django-mptt==0.7.4

# database
psycopg2==2.6

# debugging
ipython==3.1.0
ipdb==0.8.1
Werkzeug==0.10.4
